remove need for final dedent. code aligned during rendering
