+++ docs/language/line_continuations.md
@@ -16,7 +16,7 @@ a markdown paragraph can continues\
 ```python
 """a markdown paragraph can continues"""\
 \
-\
+        \
 .upper()
 ```
